# api/openapi.yaml
openapi: 3.0.3
info:
  title: TradingCopilot API
  version: 0.1.0
  description: JSON:API contract target for the Go modular monolith migration.
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /logs:
    get:
      operationId: getLogs
      parameters:
        - $ref: "#/components/parameters/LogFrom"
        - $ref: "#/components/parameters/LogTo"
        - $ref: "#/components/parameters/LogLevel"
        - $ref: "#/components/parameters/LogRole"
        - $ref: "#/components/parameters/LogSource"
        - $ref: "#/components/parameters/LogEvent"
        - $ref: "#/components/parameters/LogGroup"
        - $ref: "#/components/parameters/LogMethod"
        - $ref: "#/components/parameters/LogPath"
        - $ref: "#/components/parameters/LogStatus"
        - $ref: "#/components/parameters/LogSlowOnly"
        - $ref: "#/components/parameters/LogIncludeNoise"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/LogFile"
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200": { $ref: "#/components/responses/LogEntryCollectionDocument" }
        "400": { $ref: "#/components/responses/JsonApiErrorDocument" }
...
  /message-subscriptions:
    get:
      operationId: getMessageSubscriptions
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200": { $ref: "#/components/responses/MessageSubscriptionCollectionDocument" }
    post:
      operationId: postMessageSubscriptions
      requestBody: { $ref: "#/components/requestBodies/MessageSubscriptionUpsertRequest" }
      responses:
        "200": { $ref: "#/components/responses/MessageSubscriptionDocument" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  requestBodies:
...
    MessageSubscriptionUpsertRequest:
      required: true
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MessageSubscriptionUpsertDocument"
...
  responses:
...
    LogEntryCollectionDocument:
      description: Log entry JSON:API collection.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/LogEntryCollectionDocument"
...
    JsonApiErrorDocument:
      description: JSON:API error document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/JsonApiErrorDocument"
...
    MessageSubscriptionCollectionDocument:
      description: Message subscription JSON:API collection.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MessageSubscriptionCollectionDocument"
...
    MessageSubscriptionDocument:
      description: Message subscription JSON:API document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MessageSubscriptionDocument"
...
  parameters:
    SearchQuery:
      name: q
      in: query
      schema: { type: string }
...
    LogFrom:
      name: from
      in: query
      schema: { type: string, format: date-time }
    LogTo:
      name: to
      in: query
      schema: { type: string, format: date-time }
    LogLevel:
      name: level
      in: query
      schema:
        type: string
        enum: [debug, info, warn, error]
    LogRole:
      name: role
      in: query
      schema: { type: string }
    LogSource:
      name: source
      in: query
      schema: { type: string }
    LogEvent:
      name: event
      in: query
      schema:
        type: string
    LogGroup:
      name: group
      in: query
      schema: { type: string }
    LogMethod:
      name: method
      in: query
      schema: { type: string }
    LogPath:
      name: path
      in: query
      schema: { type: string }
    LogStatus:
      name: status
      in: query
      schema: { type: string }
    LogSlowOnly:
      name: slowOnly
      in: query
      schema: { type: boolean }
    LogIncludeNoise:
      name: includeNoise
      in: query
      schema: { type: boolean, default: false }
    LogFile:
      name: file
      in: query
      schema: { type: string }
...
    PageLimit:
      name: page[limit]
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    PageCursor:
      name: page[cursor]
      in: query
      schema:
        type: string
...
  schemas:
...
    JsonApiErrorDocument:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ErrorObject"
...
    LogEntryCollectionDocument:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/LogEntryResource"
        links:
          $ref: "#/components/schemas/Links"
        meta:
          type: object
          additionalProperties: true
    LogEntryResource:
      type: object
      required: [type, id, attributes]
      properties:
        type:
          type: string
          enum: [log-entries]
        id:
          type: string
        attributes:
          $ref: "#/components/schemas/LogEntryAttributes"
    LogEntryAttributes:
      type: object
      required: [level, role, source, event, message, caller, file, group, method, path, status, durationMs, noise, fields, raw]
      properties:
        time:
          type: string
          format: date-time
          nullable: true
        level:
          type: string
        role:
          type: string
        source:
          type: string
        event:
          type: string
        message:
          type: string
        caller:
          type: string
        file:
          type: string
        group:
          type: string
        method:
          type: string
        path:
          type: string
        status:
          type: string
        durationMs:
          type: integer
          nullable: true
        noise:
          type: boolean
        fields:
          type: object
          additionalProperties: true
        raw:
          type: object
          additionalProperties: true
        stacktrace:
          type: string
...
    MessageSubscriptionCollectionDocument:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MessageSubscriptionResource"
    MessageSubscriptionDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/MessageSubscriptionResource"
...
    MessageSubscriptionResource:
      type: object
      required: [type, id, attributes]
      properties:
        type: { type: string, enum: [message-subscriptions] }
        id: { type: string }
        attributes:
          type: object
          properties:
            provider:
              type: string
              enum: [telegram_channel, rss_feed]
            title: { type: string }
            sourceRef:
              type: string
              description: RSS/Atom source refs must be http or https feed URLs without URL-embedded credentials; use rssAuthType, rssUsername, and rssPassword for private feeds.
            enabled: { type: boolean }
            filterId: { type: integer }
            filterName: { type: string }
            teamIds:
              type: array
              items: { type: integer }
            backfillLimit: { type: integer }
            pollIntervalSeconds: { type: integer }
            collectFrom: { type: string, format: date-time }
            lastCollectedAt:
              type: string
              format: date-time
              nullable: true
            nextCollectAt:
              type: string
              format: date-time
              nullable: true
            lastCollectError:
              type: string
              nullable: true
            config:
              type: object
              additionalProperties: true
            rssAuthType:
              type: string
              enum: [none, basic, bearer]
            rssUsername: { type: string }
            hasRssPassword: { type: boolean }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    MessageSubscriptionUpsertDocument:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [type, attributes]
          properties:
            type: { type: string, enum: [message-subscriptions] }
            id: { type: string }
            attributes:
              type: object
              required: [title, sourceRef]
              properties:
                provider:
                  type: string
                  enum: [telegram_channel, rss_feed]
                title: { type: string }
                sourceRef:
                  type: string
                  description: RSS/Atom source refs must be http or https feed URLs without URL-embedded credentials; use rssAuthType, rssUsername, and rssPassword for private feeds.
                enabled: { type: boolean }
                filterId: { type: integer }
                teamIds:
                  type: array
                  items: { type: integer }
                backfillLimit: { type: integer }
                pollIntervalSeconds: { type: integer }
                config:
                  type: object
                  additionalProperties: true
                rssAuthType:
                  type: string
                  enum: [none, basic, bearer]
                rssUsername:
                  type: string
                  description: Username for RSS Basic auth. Ignored for Bearer unless a provider needs a label.
                rssPassword:
                  type: string
                  description: RSS Basic password or Bearer token. Stored as an encrypted secret and never returned.
...
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        detail:
          type: string
        field:
          type: string
    Links:
      type: object
      properties:
        self:
          type: string
        next:
          type: string
        prev:
          type: string